backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
pydantic-settings
python-jose[cryptography]
python-multipart
python-dotenv
email-validator
pytest
pytest-asyncio
httpx
